Unify Procedures/Routes rendering and fix related correctness bugs
Replaces the parallel ProcedurePoints + RoutePoints render trees with a
single Overlay-based pipeline (lib/mapGeometry, lib/overlay,
components/AviationOverlayLayers). Per-airport procedure fetch + cache
moves to procedureApi.loadAirportProcedures so the sidebar and route
resolver hit each navdata airport at most once. Drops the duplicate
procedureKey, the duplicate isTrunkTransition / isRouteRelevantSequence,
the redundant LATLON_SHORT_RE, the unreachable splitDotToken STAR
branch, and the dead emptyRoute / errorRoute helpers. Extracts
MenuDropdown to share the open/close + outside-click skeleton between
RouteDropdown and ProceduresDropdown.

Also fixes a few latent bugs the review surfaced:
- handleRouteSubmit now uses an epoch counter so a stale buildRoute can't
  clobber a newer one.
- routeBuilder body iteration wraps each token resolution in try/catch
  and pushes per-token failures into route.errors instead of aborting
  the whole route on a transient 5xx.
- handleRouteSubmit zaps the previous route state before applying the
  new one so the fix sources fully unmount + remount on each submit;
  setData alone doesn't reliably refresh Mapbox symbol-layer text when
  data swaps in place.
- findAirwayPath caches its sub-segment + identifier index per Airway
  via a WeakMap so repeated lookups skip the rebuild.
- console.log('[route] ...') gated behind import.meta.env.DEV.
