+++ src/lib/procedureGeojson.ts
@@ -1,18 +1,25 @@
 import { FixDescription, LegType, Point, Procedure, ProcedureKind, Sequence } from '~/lib/types';
+import {
+  angleDifference,
+  ArrowFeature,
+  ARROW_LEGTYPES,
+  bearingBetween,
+  destinationPoint,
+  distanceBetween,
+  FixFeature,
+  interpolateBearing,
+  isValidCoord,
+  LineSegment,
+  MANUAL_ARROW_DISTANCE_NM,
+  SequenceGeometry,
+  toTrue,
+} from '~/lib/mapGeometry';
 
 const SID_SKIP_LEGTYPES: LegType[] = ['HeadingToAltitude', 'CourseToAltitude'];
-export const ARROW_LEGTYPES: LegType[] = [
-  'HeadingToManual',
-  'CourseToManual',
-  'FromFixToManual',
-  'ManualTermination',
-];
 const MISSED_APPROACH_STOPPERS: FixDescription[] = ['MissedApproach', 'MissedApproachFirstLeg'];
 
-export const MANUAL_ARROW_DISTANCE_NM = 5;
 const ARROW_WING_LENGTH_NM = 0.6;
 const ARROW_WING_ANGLE_DEG = 30;
-const EARTH_RADIUS_NM = 3440.065;
 const ARC_THRESHOLD_DEG = 60;
 const ARC_DISTANCE_NM = MANUAL_ARROW_DISTANCE_NM / 2;
 const ARC_STEPS = 12;
@@ -41,51 +48,12 @@ export const formatFixLabel = (point: Point, kind: ProcedureKind): string => {
 };
 
 const hasCoords = (p: Point): p is Point & { latitude: number; longitude: number } =>
-  p.latitude != null &&
-  p.longitude != null &&
-  Number.isFinite(p.latitude) &&
-  Number.isFinite(p.longitude) &&
-  !(p.latitude === 0 && p.longitude === 0);
+  isValidCoord(p.latitude, p.longitude);
 
 const isMissedApproachBoundary = (point: Point): boolean =>
   point.descriptions.some((d) => MISSED_APPROACH_STOPPERS.includes(d));
 
 const isRunwayThreshold = (point: Point): boolean => point.descriptions.includes('RunwayHelipad');
-
-export const destinationPoint = (lat: number, lon: number, bearingDeg: number, distanceNM: number) => {
-  const angular = distanceNM / EARTH_RADIUS_NM;
-  const bearing = (bearingDeg * Math.PI) / 180;
-  const lat1 = (lat * Math.PI) / 180;
-  const lon1 = (lon * Math.PI) / 180;
-  const lat2 = Math.asin(
-    Math.sin(lat1) * Math.cos(angular) + Math.cos(lat1) * Math.sin(angular) * Math.cos(bearing),
-  );
-  const lon2 =
-    lon1 +
-    Math.atan2(
-      Math.sin(bearing) * Math.sin(angular) * Math.cos(lat1),
-      Math.cos(angular) - Math.sin(lat1) * Math.sin(lat2),
-    );
-  return { latitude: (lat2 * 180) / Math.PI, longitude: (lon2 * 180) / Math.PI };
-};
-
-export const bearingBetween = (lat1: number, lon1: number, lat2: number, lon2: number): number => {
-  const phi1 = (lat1 * Math.PI) / 180;
-  const phi2 = (lat2 * Math.PI) / 180;
-  const dLon = ((lon2 - lon1) * Math.PI) / 180;
-  const y = Math.sin(dLon) * Math.cos(phi2);
-  const x = Math.cos(phi1) * Math.sin(phi2) - Math.sin(phi1) * Math.cos(phi2) * Math.cos(dLon);
-  return ((Math.atan2(y, x) * 180) / Math.PI + 360) % 360;
-};
-
-export const distanceBetween = (lat1: number, lon1: number, lat2: number, lon2: number): number => {
-  const phi1 = (lat1 * Math.PI) / 180;
-  const phi2 = (lat2 * Math.PI) / 180;
-  const dPhi = ((lat2 - lat1) * Math.PI) / 180;
-  const dLon = ((lon2 - lon1) * Math.PI) / 180;
-  const a = Math.sin(dPhi / 2) ** 2 + Math.cos(phi1) * Math.cos(phi2) * Math.sin(dLon / 2) ** 2;
-  return 2 * EARTH_RADIUS_NM * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-};
 
 const runwayHeading = (transition?: string): number | null => {
   if (!transition) return null;
@@ -93,21 +61,6 @@ const runwayHeading = (transition?: string): number | null => {
   if (!m) return null;
   return (parseInt(m[1], 10) * 10) % 360;
 };
-
-const angleDifference = (a: number, b: number): number =>
-  Math.abs(((b - a + 540) % 360) - 180);
-
-const interpolateBearing = (from: number, to: number, t: number): number => {
-  const diff = ((to - from + 540) % 360) - 180;
-  return (from + diff * t + 360) % 360;
-};
-
-// Magnetic→true conversion. The geometry layer (destinationPoint, bearingBetween)
-// works against geographic north, so any magnetic-domain input (point.course
-// from the API, runway-designator-derived headings) must be converted before
-// being plotted. variation > 0 = easterly.
-export const toTrue = (magneticDeg: number, correction: number): number =>
-  (magneticDeg + correction + 360) % 360;
 
 /**
  * Generate arc points that smoothly transition from rwyHeading to courseHeading
@@ -131,35 +84,6 @@ const generateArcPoints = (
   }
   return { coords, end: pos };
 };
-
-export interface FixFeature {
-  type: 'Feature';
-  geometry: { type: 'Point'; coordinates: [number, number] };
-  properties: { text: string; identifier: string };
-}
-
-export interface ArrowFeature {
-  type: 'Feature';
-  geometry: { type: 'LineString'; coordinates: [number, number][] };
-  properties: Record<string, never>;
-}
-
-export interface LineSegment {
-  coords: [number, number][];
-  dashed: boolean;
-}
-
-export interface LineFeature {
-  type: 'Feature';
-  geometry: { type: 'LineString'; coordinates: [number, number][] };
-  properties: { dashed: boolean };
-}
-
-export interface SequenceGeometry {
-  lineSegments: LineSegment[];
-  fixFeatures: FixFeature[];
-  arrowFeatures: ArrowFeature[];
-}
 
 const flushSegment = (
   segments: LineSegment[],
@@ -328,74 +252,10 @@ export const buildSequenceGeometry = (
   return { lineSegments, fixFeatures, arrowFeatures };
 };
 
-export const lineSegmentsToFeatures = (segments: LineSegment[]): LineFeature[] =>
-  segments.map((seg) => ({
-    type: 'Feature',
-    geometry: { type: 'LineString', coordinates: seg.coords },
-    properties: { dashed: seg.dashed },
-  }));
-
-export interface ProcedureSequenceGeometry {
-  sequence: Sequence;
-  lineSegments: LineSegment[];
-  arrowFeatures: ArrowFeature[];
-}
-
-export interface ProcedureGeometry {
-  sequences: ProcedureSequenceGeometry[];
-  fixFeatures: FixFeature[];
-}
-
-const pickBetterFix = (a: FixFeature, b: FixFeature): FixFeature =>
-  b.properties.text.length > a.properties.text.length ? b : a;
-
-export const buildProcedureGeometry = (procedure: Procedure): ProcedureGeometry => {
-  const sequencesGeom: ProcedureSequenceGeometry[] = [];
-  const bestByIdentifier = new Map<string, FixFeature>();
-  const correction = procedure.magneticCorrection ?? 0;
-
-  for (const sequence of procedure.sequences) {
-    const geom = buildSequenceGeometry(sequence, procedure.kind, correction);
-    sequencesGeom.push({
-      sequence,
-      lineSegments: geom.lineSegments,
-      arrowFeatures: geom.arrowFeatures,
-    });
-    for (const feat of geom.fixFeatures) {
-      const id = feat.properties.identifier;
-      if (!id) continue;
-      const existing = bestByIdentifier.get(id);
-      bestByIdentifier.set(id, existing ? pickBetterFix(existing, feat) : feat);
-    }
-  }
-
-  return { sequences: sequencesGeom, fixFeatures: Array.from(bestByIdentifier.values()) };
-};
-
 export const procedureKey = (procedure: Procedure): string =>
   `${procedure.kind}:${procedure.airport}:${procedure.identifier}`;
 
 export const sequenceLayerId = (procedure: Procedure, sequence: Sequence): string => {
   const transition = sequence.transition ? sequence.transition : 'null';
   return `${procedure.kind}-${procedure.airport}-${procedure.identifier}-${transition}`;
-};
-
-export const procedureLayerId = (procedure: Procedure): string =>
-  `${procedure.kind}-${procedure.airport}-${procedure.identifier}`;
-
-export const aggregateFixFeatures = (procedures: Procedure[]): FixFeature[] => {
-  const bestByIdentifier = new Map<string, FixFeature>();
-  for (const procedure of procedures) {
-    const correction = procedure.magneticCorrection ?? 0;
-    for (const sequence of procedure.sequences) {
-      const geom = buildSequenceGeometry(sequence, procedure.kind, correction);
-      for (const feat of geom.fixFeatures) {
-        const id = feat.properties.identifier;
-        if (!id) continue;
-        const existing = bestByIdentifier.get(id);
-        bestByIdentifier.set(id, existing ? pickBetterFix(existing, feat) : feat);
-      }
-    }
-  }
-  return Array.from(bestByIdentifier.values());
 };
